# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Manageemnt API",
        "description": "Library Manageemnt API for Swagger integration",
        "contact": {
            "name": "Rajan Bhatta (+88 0162119979, aobak63@gmail.com)",
            "email": "aobak63@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get all books",
                "operationId": "64ce85f06fc87443a16496268e88cddb",
                "responses": {
                    "200": {
                        "description": "List of books"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "operationId": "4114213d13c3b8499c6aec21511edb28",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "published_year",
                                    "bookshelf_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "published_year": {
                                        "type": "integer"
                                    },
                                    "bookshelf_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a single book",
                "operationId": "e473ab3a66f0daaa9bf1b70b29dd5288",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book details"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update a book",
                "operationId": "7341e0ef7ae7c978e0f7ef257d9dbba9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "published_year": {
                                        "type": "integer"
                                    },
                                    "bookshelf_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book",
                "operationId": "b74ac8316831ae82112b04392815bd67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book deleted"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            }
        },
        "/api/books/search": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Search for books",
                "operationId": "a8bcea204ef320eb8d8bf51f66a45089",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "The search keyword for books",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results"
                    }
                },
                "security": [
                    {
                        "ApiTokenAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiTokenAuth": {
                "type": "apiKey",
                "description": "Token based authentication using X-API-TOKEN header",
                "name": "X-API-TOKEN",
                "in": "header"
            }
        }
    },
...
}